got progress calculation working on more pages. Including map pages.
