+++ map.html
@@ -4,7 +4,6 @@
 	<link rel="icon" type="image/png" href="./favicon.ico"/>
 	<link rel="stylesheet" type="text/css" href="main.css"/>
 	<link rel="stylesheet" type="text/css" href="guide.css"/>
-	<script src="sharing.js"></script>
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>Map — Interactive Oblivion Checklist</title>
 </head>
@@ -123,6 +122,8 @@
 	import * as map from './js/map.mjs';
 	import * as point from './js/map/point.mjs';
 	import * as obliviondata from './js/obliviondata.mjs';
+	import * as sharing from './js/sharing.mjs';
+	import * as progress from './js/progressCalculation.mjs';
 	
 	if(userdata.loadCookie("debug") == true){
 		window.debug = true;
@@ -166,6 +167,7 @@
 			//need to add events to redraw immediately upon change
 			//TODO fix. or move all this stuff into another js file.  or something.
 			document.getElementById("mapShowPrediscovered").addEventListener('change',()=>{map.draw();});
+			progress.recalculateProgress();
 		});
 	}).call(window);
